+++ CHANGELOG.md
@@ -13,6 +13,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
   - CTRL+T: Expand Talkers
   - CTRL+SHIFT+T: Collapse Talkers
 - Connection matrix displays Entity Connection Summary (squares) using color code (see Legend screen for more details)
+- Option to collapse all connection matrix headers for new entities
 
 ## [1.2.8] - 2022-04-26
 ### Added